Move util package from internal to pkg so it can be used to generate data for verification server tests.
